filter.validate_arguments: Move argument validation to here
Also, validate arguments in the top-level `filter.run`. This allows the
imported run() function to be responsible for one less thing.
